Combine functionality into single program
Make concat accessible via -export <exportType> flag
Make export_to_json accessible via -import flag
Update README.md accordingly
Update Github Workflow to use -export option
